+++ src/main/java/wanted/ribbon/datapipe/config/AppConfig.java
@@ -1,4 +1,4 @@
-package wanted.ribbon.genrestrt.config;
+package wanted.ribbon.datapipe.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.context.annotation.Bean;
